Rename ChannelTypes -> ChannelData
And move commands to a new ChannelCommands file.
This commit doesn't contain any logic changes.
